backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
boto3==1.35.76
pydantic-settings==2.6.1
python-dotenv==1.0.1
requests==2.32.3